+++ .project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>mmv-tracking-napari</name>
+	<name>mmv_hitl4trk</name>
 	<comment></comment>
 	<projects>
 	</projects>
