+++ LICENSE
@@ -12,7 +12,7 @@ modification, are permitted provided that the following conditions are met:
   this list of conditions and the following disclaimer in the documentation
   and/or other materials provided with the distribution.
 
-* Neither the name of mmv-tracking-napari nor the names of its
+* Neither the name of mmv_hitl4trk nor the names of its
   contributors may be used to endorse or promote products derived from
   this software without specific prior written permission.
