+++ mkdocs.yml
@@ -1,11 +1,11 @@
-site_name: mmv-tracking-napari
+site_name: mmv-HITL_4TRK
 site_description: A simple plugin to use with napari
 site_author: lennart kowitz
 
 theme: readthedocs
 
 
-repo_url: https://github.com/MMV-Lab/mmv-tracking-napari
+repo_url: https://github.com/MMV-Lab/mmv_hitl4trk
 
 pages:
 - Home: index.md
