+++ setup.cfg
@@ -1,9 +1,9 @@
 [metadata]
-name = mmv-tracking-napari
+name = mmv_hitl4trk
 version = 1.0.0
 author = lennart kowitz
 author_email = lennart.kowitz@isas.de
-url = https://github.com/MMV-Lab/mmv-tracking-napari
+url = https://github.com/MMV-Lab/mmv_hitl4trk
 license = BSD-3-Clause
 description = A simple plugin to use with napari
 long_description = file: README.md
@@ -22,10 +22,10 @@ classifiers =
     Operating System :: OS Independent
     License :: OSI Approved :: BSD License
 project_urls =
-    Bug Tracker = https://github.com/MMV-Lab/mmv-tracking-napari/issues
-    Documentation = https://github.com/MMV-Lab/mmv-tracking-napari#README.md
-    Source Code = https://github.com/MMV-Lab/mmv-tracking-napari
-    User Support = https://github.com/MMV-Lab/mmv-tracking-napari/issues
+    Bug Tracker = https://github.com/MMV-Lab/mmv_hitl4trk/issues
+    Documentation = https://github.com/MMV-Lab/mmv_hitl4trk#README.md
+    Source Code = https://github.com/MMV-Lab/mmv_hitl4trk
+    User Support = https://github.com/MMV-Lab/mmv_hitl4trk/issues
 
 [options]
 packages = find:
@@ -54,8 +54,8 @@ where = src
 
 [options.entry_points] 
 napari.manifest = 
-    mmv-tracking-napari = mmv_tracking_napari:napari.yaml
+    mmv_hitl4trk = mmv_hitl4trk:napari.yaml
 
 [options.package_data]
-mmv_tracking_napari =
+mmv_hitl4trk =
     napari.yaml
