+++ src/mmv_hitl4trk/__init__.py
@@ -8,6 +8,6 @@ IOU_THRESHOLD = .4
 APPROX_INF = 65535
 MAX_MATCHING_DIST = 45
 
-from ._widget import MMVTracking
+from ._widget import MMVHITL4TRK
 
-__all__ = ["MMVTracking"]
+__all__ = ["MMVHITL4TRK"]
