+++ src/mmv_hitl4trk/_tests/data/images/Raw Image.tif
