+++ src/mmv_hitl4trk/_tests/data/segmentation/GT.tif
