+++ src/mmv_hitl4trk/_tests/data/segmentation/GT_false_negative_2.tif
