+++ src/mmv_hitl4trk/_tests/data/segmentation/centroid_ausserhalb.tif
