+++ src/mmv_hitl4trk/_tests/data/segmentation/false positive.tif
