+++ src/mmv_hitl4trk/_tests/data/segmentation/false_negative_1.tif
