+++ src/mmv_hitl4trk/_tests/data/segmentation/falsely_cut.tif
