+++ src/mmv_hitl4trk/_tests/data/segmentation/zelle_kleiner.tif
