+++ src/mmv_hitl4trk/_tests/data/tracks/GT_tracks.npy
