+++ src/mmv_hitl4trk/_tests/data/tracks/centroid_outside.npy
