+++ src/mmv_hitl4trk/_tests/data/tracks/deleted_edge.npy
