+++ src/mmv_hitl4trk/_tests/data/tracks/switch.npy
