--- src/mmv_tracking_napari/napari.yaml
@@ -1,11 +1,0 @@
-name: mmv-tracking-napari
-display_name: Cell Tracking Tool
-schema_version: 0.1.0
-contributions:
-  commands: 
-  - id: mmv-tracking-napari.make_qwidget
-    python_name: mmv_tracking_napari._widget:MMVTracking
-    title: Make tracking widget
-  widgets:
-  - command: mmv-tracking-napari.make_qwidget
-    display_name: Tracking Widget
