+++ tox.ini
@@ -34,4 +34,4 @@ deps =
     pytest-qt
     qtpy
     pyqt5
-commands = pytest -v --color=yes --cov=mmv_tracking_napari --cov-report=xml
+commands = pytest -v --color=yes --cov=mmv_hitl4trk --cov-report=xml
